+++ src/lib/components/LogViewer.svelte
